# openapi.yaml
openapi: 3.0.0
info:
  title: Budget Management API
  description: A comprehensive backend API for managing budgets, expenses, orders, transactions, and notifications.
  version: 1.0.0
  contact:
    name: Son Nguyen
    url: https://sonnguyenhoang.com
    email: hoangson091104@gmail.com
servers:
  - url: https://mongo-redis-rabbitmq-kafka-elk-backend.onrender.com
    description: Production Server
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      required:
        - customerId
        - amount
      properties:
        _id:
          type: string
          description: The unique identifier for the order.
        customerId:
          type: string
          description: The ID of the associated customer.
        amount:
          type: number
          description: The total amount for the order.
        status:
          type: string
          description: The status of the order.
          default: pending
        createdAt:
          type: string
          format: date-time
          description: When the order was created.
    Task:
      type: object
      required:
        - description
      properties:
        _id:
          type: string
          description: The unique identifier for the task.
        description:
          type: string
          description: Details about the task.
        status:
          type: string
          description: The status of the task.
          default: pending
        createdAt:
          type: string
          format: date-time
          description: When the task was created.
paths:
...
  /api/orders:
    post:
      summary: Create a new order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        201:
          description: Order created successfully.
    get:
      summary: Retrieve all orders
      tags: [Orders]
      responses:
        200:
          description: List of orders.
  /api/orders/{id}:
    get:
      summary: Retrieve an order by ID
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order retrieved successfully.
    delete:
      summary: Delete an order by ID
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order deleted successfully.
  /api/expenses/search:
    post:
      summary: Search expenses in Elasticsearch
      tags: [Search]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
      responses:
        200:
          description: Search results.
  /api/tasks:
    post:
      summary: Submit a task to RabbitMQ
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        202:
          description: Task submitted successfully.
  /api/tasks/{id}:
    get:
      summary: Check the status of a task
      tags: [Tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Task status.
  /api/transactions:
    post:
      summary: Add a new transaction log
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                description:
                  type: string
                amount:
                  type: number
                budgetId:
                  type: string
      responses:
        201:
          description: Transaction added successfully.
    get:
      summary: Retrieve all transaction logs
      tags: [Transactions]
      responses:
        200:
          description: List of transactions.
  /api/transactions/user/{userId}:
    get:
      summary: Retrieve transaction logs by user ID
      tags: [Transactions]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Transaction logs for the user.
  /api/users/profile:
    get:
      summary: Get the profile of the currently authenticated user
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        200:
          description: User profile retrieved successfully.
    put:
      summary: Update the profile of the currently authenticated user
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
      responses:
        200:
          description: User profile updated successfully.